+++ spacerocks/spacerock.py
@@ -9,7 +9,6 @@ import datetime
 import numpy as np
 import pandas as pd
 
-import rebound
 import asdf
 import copy
 
@@ -21,24 +20,10 @@ from .vector import Vector
 from .ephemerides import Ephemerides 
 from .observer import Observer
 from .cbindings import kepM_to_xyz, correct_for_ltt
-from .spice import SpiceBody
-import os
+from .spice import SpiceBody, SpiceKernel
+#from .simulation import Simulation
 import pkg_resources
-import spiceypy as spice
-
-
 SPICE_PATH = pkg_resources.resource_filename('spacerocks', 'data/spice')
-spice.furnsh(os.path.join(SPICE_PATH, 'latest_leapseconds.tls'))
-#spice.furnsh(os.path.join(SPICE_PATH, 'de440s.bsp'))
-#spice.furnsh(os.path.join(SPICE_PATH, 'de441.bsp'))
-#spice.furnsh(os.path.join(SPICE_PATH, 'hst.bsp'))
-#spice.furnsh(os.path.join(SPICE_PATH, 'nh_pred_alleph_od151.bsp'))
-spice.furnsh(os.path.join(SPICE_PATH, 'nh_pred_alleph_od151.bsp'))
-spice.furnsh(os.path.join(SPICE_PATH, 'de423.bsp'))
-
-sun = SpiceBody(spiceid='Sun')
-earth = SpiceBody(spiceid='Earth')
-
 DATA_PATH = pkg_resources.resource_filename('spacerocks', 'data/observatories.csv')
 observatories = pd.read_csv(DATA_PATH)
 
@@ -49,8 +34,10 @@ class SpaceRock(KeplerOrbit, Convenience):
     '''
     def __init__(self, origin='ssb', frame='eclipJ2000', units=Units(), *args, **kwargs):
 
+        kernel = SpiceKernel()
+        kernel.furnsh()
+
         coords = self.detect_coords(kwargs)
-        #frame = frame.lower()
         origin = origin.lower()
 
         # input -> arrays
@@ -75,6 +62,7 @@ class SpaceRock(KeplerOrbit, Convenience):
             o = SpiceBody(spiceid=self.origin)
             self.mu = o.mu
 
+        self.units = units
         if coords == 'kep':
 
             if kwargs.get('a') is not None:
@@ -155,7 +143,6 @@ class SpaceRock(KeplerOrbit, Convenience):
                 self.name = np.atleast_1d(kwargs.get('name'))
             else:
                 # produces random, non-repeting integers between 0 and 1e10 - 1
-                #self.name = np.array(['{:010}'.format(value) for value in random.sample(range(int(1e10)), len(self.inc))])
                 self.name = np.array([uuid.uuid4().hex for _ in range(len(self.inc))])
 
 
@@ -183,8 +170,7 @@ class SpaceRock(KeplerOrbit, Convenience):
             if kwargs.get('name') is not None:
                 self.name = np.atleast_1d(kwargs.get('name'))
             else:
-                # produces random, non-repeting integers between 0 and 1e10 - 1
-                #self.name = np.array(['{:010}'.format(value) for value in random.sample(range(int(1e10)), len(self.x))])
+                # produces random, non-repeting identifiers
                 self.name = np.array([uuid.uuid4().hex for _ in range(len(self.x))])
 
         
@@ -238,6 +224,33 @@ class SpaceRock(KeplerOrbit, Convenience):
         if kwargs.get('density') is not None:
             self.density = kwargs.get('density') * units.density
 
+        if kwargs.get('cov') is not None:
+            self.cov = kwargs.get('cov')
+
+
+
+    def clone(self, N: int):
+        epochs = np.repeat(self.epoch.tdb.jd, N)
+        names = [self.name[0] + f'-clone{idx}' for idx in range(N)]
+
+        mean = [self.x.au[0], 
+                self.y.au[0], 
+                self.z.au[0], 
+                self.vx.value[0], 
+                self.vy.value[0], 
+                self.vz.value[0]]
+        xs, ys, zs, vxs, vys, vzs = np.random.multivariate_normal(mean, self.cov[0], N).T
+    
+        # units = Units()
+        # units.distance = u.au
+        # units.speed = u.au/u.day
+        # units.timescale = 'tdb'
+
+        return SpaceRock(x=xs, y=ys, z=zs, vx=vxs, vy=vys, vz=vzs, name=names,
+                         origin=self.origin, frame=self.frame, units=self.units, epoch=epochs)
+
+
+
 
     @classmethod
     def from_horizons(cls, name):
@@ -245,7 +258,6 @@ class SpaceRock(KeplerOrbit, Convenience):
         from urllib.parse import urlencode
         from urllib.request import urlopen
        
-
         def quote(text):
             return "'{}'".format(text)
 
@@ -262,6 +274,7 @@ class SpaceRock(KeplerOrbit, Convenience):
         Gs = []
         epochs = []
         names = []
+        covs = []
 
         for n in np.atleast_1d(name):
         
@@ -277,10 +290,10 @@ class SpaceRock(KeplerOrbit, Convenience):
                       'STEP_SIZE':   quote("1"),
                       'REF_SYSTEM':  quote("J2000"),
                       'VEC_CORR':    quote("NONE"),
-                      'OUT_UNITS':   quote("KM-S"),
-                      'CSV_FORMAT':  quote("NO"),
+                      'OUT_UNITS':   quote("AU-D"),
+                      'CSV_FORMAT':  quote("YES"),
                       'VEC_DELTA_T': quote("NO"),
-                      'VEC_TABLE':   quote("3"),
+                      'VEC_TABLE':   quote("2x"),
                       'VEC_LABELS':  quote("NO")
                      }
     
@@ -290,19 +303,23 @@ class SpaceRock(KeplerOrbit, Convenience):
                     body = f.read().decode()
 
                 lines = body.split("$$SOE")[-1].split("\n")
-                x, y, z = [float(i) for i in lines[2].split()]
-                vx, vy, vz = [float(i) for i in lines[3].split()]
+                spl = [float(i) for i in lines[2].split(',')[2:-1]]
+                epoch = float(lines[2].split(',')[0])
+                x, y, z, vx, vy, vz = spl[:6]
+                dx, dy, dz, dvx, dvy, dvz = spl[6:]
+
+                cov = np.diag([dx**2, dy**2, dz**2, dvx**2, dvy**2, dvz**2])
         
                 try:
                     pps = body.split('physical parameters')[-1].split('\n')[2].split()
                     H = float(pps[1])
                     G = float(pps[3])
                 except Exception as E:
-                    print(f'Magnitude information not available for {name}. Is this a comet?')
+                    print(f'Magnitude information not available for {name}. Is this a comet? Or a planet?')
                     H = None
                     G = None
                     pass
-    
+
                 xs.append(x)
                 ys.append(y)
                 zs.append(z)
@@ -311,19 +328,20 @@ class SpaceRock(KeplerOrbit, Convenience):
                 vzs.append(vz)
                 Hs.append(H)
                 Gs.append(G)
-                epochs.append(start.ctime())
+                epochs.append(epoch)
                 names.append(n)
+                covs.append(cov)
 
             except Exception as E:
                 raise ValueError(f'Body {n} not found.')
             
         units = Units()
-        units.distance = u.km
-        units.speed = u.km/u.s
+        units.distance = u.au
+        units.speed = u.au/u.day
         units.timescale = 'tdb'
 
         return cls(x=xs, y=ys, z=zs, vx=vxs, vy=vys, vz=vzs, name=names,
-                   origin='sun', frame='eclipJ2000', units=units, H=Hs, G=Gs, epoch=epochs)
+                   origin='sun', frame='eclipJ2000', units=units, H=Hs, G=Gs, epoch=epochs, cov=covs)
         
     @classmethod
     def from_file(cls, name):
@@ -363,12 +381,9 @@ class SpaceRock(KeplerOrbit, Convenience):
         if hasattr(self, 'mag_func'):
             return np.array([func(epoch) for epoch, func in zip(self.epoch.jd, self.mag_func)])
 
-    def propagate(self, epochs, model, units=Units(), gr=False):
-        '''
-        Numerically integrate all bodies to the desired date.
-        This routine synchronizes the epochs.
-        '''
 
+    def propagate(self, epochs, model='GIANTS', units=Units(), gr=False):
+        from .simulation import Simulation
         epochs = self.detect_timescale(np.atleast_1d(epochs), units.timescale)
         origin = copy.copy(self.origin)
         frame = copy.copy(self.frame)
@@ -379,110 +394,61 @@ class SpaceRock(KeplerOrbit, Convenience):
         # Put the simulation into ecliptic coordinates
         self.change_frame('eclipJ2000')
 
-        # Integrate all particles to the same obsdate
-        pickup_times = self.epoch.tdb.jd
-        sim, planet_names = self.set_simulation(np.min(pickup_times), model=model, gr=gr)
-        sim.t = np.min(pickup_times)
+        units = Units()
+        units.timescale = 'tdb'
+        units.timeformat = 'jd'
+        sim = Simulation(model=model, epoch=self.epoch.tdb.jd, units=units)
+        sim.add_spacerocks(self)
+        sim.N_active = len(sim.model.perturbers)
+        sim.testparticle_type = 0
+        sim.integrator = 'ias15'
 
-        # need to ensure these are computed
-        self.x
-        self.y
-        self.z
-        self.vx
-        self.vy
-        self.vz
-
-        sim.move_to_com()
-
-        for time in np.sort(np.unique(pickup_times)):
-            ps = self[self.epoch.tdb.jd == time]
-            for x, y, z, vx, vy, vz, name in zip(ps.x.value, ps.y.value, ps.z.value, ps.vx.value, ps.vy.value, ps.vz.value, ps.name):
-                sim.add(x=x, y=y, z=z, vx=vx, vy=vy, vz=vz, hash=name)
-                sim.integrate(time, exact_finish_time=1)
-
-        Nx = len(epochs)
-        x_values = np.zeros((Nx, sim.N))
-        y_values = np.zeros((Nx, sim.N))
-        z_values = np.zeros((Nx, sim.N))
-        vx_values = np.zeros((Nx, sim.N))
-        vy_values = np.zeros((Nx, sim.N))
-        vz_values = np.zeros((Nx, sim.N))
-        name_values = np.zeros((Nx, sim.N), dtype='object')
-        obsdate_values = np.zeros((Nx, sim.N))
-
-        p_names = planet_names + [name for name in self.name]
-
-        for ii, time in enumerate(np.sort(epochs.tdb.jd)):
-            sim.integrate(time, exact_finish_time=1)
-            a = np.zeros((sim.N, 3), dtype='float64')
-            b = np.zeros((sim.N, 3), dtype='float64')
-            sim.serialize_particle_data(xyz=a, vxvyvz=b)
-
-            x, y, z = a.T
-            vx, vy, vz = b.T
-
-            x_values[ii] = x
-            y_values[ii] = y
-            z_values[ii] = z
-            vx_values[ii] = vx
-            vy_values[ii] = vy
-            vz_values[ii] = vz
-            name_values[ii] = p_names
-            obsdate_values[ii] = np.repeat(sim.t, sim.N)
-
-        Nactive = sim.N_active
-        x = x_values[:, Nactive:].flatten()
-        y = y_values[:, Nactive:].flatten()
-        z = z_values[:, Nactive:].flatten()
-        vx = vx_values[:, Nactive:].flatten()
-        vy = vy_values[:, Nactive:].flatten()
-        vz = vz_values[:, Nactive:].flatten()
-        name = name_values[:, Nactive:].flatten()
-        epoch = obsdate_values[:, Nactive:].flatten()
-
-        px = x_values[:, :Nactive].flatten()
-        py = y_values[:, :Nactive].flatten()
-        pz = z_values[:, :Nactive].flatten()
-        pvx = vx_values[:, :Nactive].flatten()
-        pvy = vy_values[:, :Nactive].flatten()
-        pvz = vz_values[:, :Nactive].flatten()
-        pname = name_values[:, :Nactive].flatten()
-        pepoch = obsdate_values[:, :Nactive].flatten()
+        if gr == True:
+            import reboundx
+            rebx = reboundx.Extras(sim)
+            gr_force = rebx.load_force('gr_full')
+            gr_force.params["c"] = c.to(u.au / u.day).value
+            rebx.add_force(gr_force)
 
         units = Units()
         units.timescale = 'tdb'
-        rocks = self.__class__(x=x, y=y, z=z, vx=vx, vy=vy, vz=vz, name=name, epoch=epoch, origin='ssb', frame='eclipJ2000', units=units)
-        planets = self.__class__(x=px, y=py, z=pz, vx=pvx, vy=pvy, vz=pvz, name=pname, epoch=pepoch, origin='ssb', frame='eclipJ2000', units=units)
+        units.timeformat = 'jd'
+        prop, planets, sim = sim.propagate(epochs=epochs.tdb.jd, units=units)
 
         # be polite and return orbital parameters using the input origin.
         if origin != 'ssb':
             self.change_origin(origin)
-            rocks.change_origin(origin)
+            prop.change_origin(origin)
             planets.change_origin(origin)
 
         if frame != 'eclipJ2000':
             self.change_frame(frame)
-            rocks.change_frame(frame)
+            prop.change_frame(frame)
             planets.change_frame(frame)
 
+        N = len(epochs)
         if hasattr(self, 'G'):
-            rocks.G = np.tile(self.G, Nx)
+            prop.G = np.tile(self.G, N)
 
         if hasattr(self, 'H_func'):
-            rocks.H_func = np.tile(self.H_func, Nx)
+            prop.H_func = np.tile(self.H_func, N)
         elif hasattr(self, '_H'):
-            rocks.H = np.tile(self.H, Nx)
+            prop.H = np.tile(self.H, N)
 
         if hasattr(self, 'mag_func'):
-            rocks.mag_func = np.tile(self.mag_func, Nx)
+            prop.mag_func = np.tile(self.mag_func, N)
 
-        return rocks, planets, sim
+        return prop, planets, sim
+
 
     def calc_H(self, obscode):
         return self.__calc_H_from_mag(obscode=obscode)
 
     def __calc_H_from_mag(self, obscode):
         obs = self.observe(obscode=obscode)
+
+        earth = SpiceBody(spiceid='Earth')
+        sun = SpiceBody(spiceid='Sun')
 
         e = earth.at(self.epoch)
         s = sun.at(self.epoch)
@@ -512,7 +478,6 @@ class SpaceRock(KeplerOrbit, Convenience):
 
 
     def observe(self, **kwargs) -> Ephemerides:
-
         '''
         Calculate the ephemerides of a SpaceRock object from a observer's location.
 
@@ -525,14 +490,15 @@ class SpaceRock(KeplerOrbit, Convenience):
         Ephemerides can also be computed as observed from the Sun and the 
         solar system barycenter (ssb).
 
-        The James Webb Space Telescope will be supported as soon as it launches
-        and NASA provides the necessary spk files.
+        The James Webb Space Telescope will be supported as soon as
+        NASA provides the necessary spk files.
         '''
-
         if kwargs.get('obscode') is not None:
             x, y, z, vx, vy, vz = self.xyz_to_tel(obscode=kwargs.get('obscode'))
         elif kwargs.get('spiceid') is not None:
             x, y, z, vx, vy, vz = self.xyz_to_tel(spiceid=kwargs.get('spiceid'))
+        elif kwargs.get('observer') is not None:
+            x, y, z, vx, vy, vz = self.xyz_to_tel(observer=kwargs.get('observer'))
         else:
             raise ValueError('Must pass either an obscode or spiceid.')
 
@@ -552,13 +518,14 @@ class SpaceRock(KeplerOrbit, Convenience):
 
         Routine corrects iteratively for light travel time.
         '''
-
         if kwargs.get('obscode') is not None:
             observer = Observer(obscode=kwargs.get('obscode'), epoch=self.epoch.utc.jd)
         elif kwargs.get('spiceid') is not None:
             observer = Observer(spiceid=kwargs.get('spiceid'), epoch=self.epoch.utc.jd)
+        elif kwargs.get('observer') is not None:
+            observer = kwargs.get('observer')
         else:
-            raise ValueError('Must pass either an obscode or spiceid.')
+            raise ValueError('Must pass either an obscode, spiceid, or Observer object.')
 
         in_origin = copy.copy(self.origin)
         self.to_bary()
@@ -588,165 +555,18 @@ class SpaceRock(KeplerOrbit, Convenience):
 
         return dx, dy, dz, dvx, dvy, dvz
 
-    def set_simulation(self, startdate, model, gr):
-
-        sun = SpiceBody(spiceid='Sun')
-        mercury = SpiceBody(spiceid='Mercury Barycenter')
-        venus = SpiceBody(spiceid='Venus Barycenter')
-        earth = SpiceBody(spiceid='Earth')
-        moon = SpiceBody(spiceid='Moon')
-        mars = SpiceBody(spiceid='Mars Barycenter')
-        jupiter = SpiceBody(spiceid='Jupiter Barycenter')
-        saturn = SpiceBody(spiceid='Saturn Barycenter')
-        uranus = SpiceBody(spiceid='Uranus Barycenter')
-        neptune = SpiceBody(spiceid='Neptune Barycenter')
-        pluto = SpiceBody(spiceid='Pluto Barycenter')
-
-        M_sun = sun.mass.value
-        M_mercury = mercury.mass.value
-        M_venus = venus.mass.value
-        M_earth = earth.mass.value
-        M_moon = moon.mass.value
-        M_mars = mars.mass.value
-        M_jupiter = jupiter.mass.value
-        M_saturn = saturn.mass.value
-        M_uranus = uranus.mass.value
-        M_neptune = neptune.mass.value
-        M_pluto = pluto.mass.value
-
-        if model == 0:
-            active_bodies = [sun]
-            names = ['Sun']
-            M_sun += M_mercury + M_venus + M_earth + M_mars \
-                     + M_jupiter + M_saturn + M_uranus + M_neptune
-            masses = [M_sun]
-
-        elif model == 1:
-            active_bodies = [sun, jupiter, saturn, uranus, neptune]
-            names = ['Sun', 'Jupiter', 'Saturn', 'Uranus', 'Neptune']
-            M_sun += M_mercury + M_venus + M_earth + M_mars
-            masses = [M_sun, M_jupiter, M_saturn, M_uranus, M_neptune]
-
-
-        elif model == 2:
-            active_bodies = [sun, mercury, venus, earth, moon, mars, jupiter, saturn, uranus, neptune, pluto]
-            names = ['Sun', 'Mercury', 'Venus', 'Earth', 'Moon', 'Mars', 'Jupiter', 'Saturn', 'Uranus', 'Neptune', 'Pluto']
-            masses = [M_sun, M_mercury, M_venus, M_earth, M_moon, M_mars, M_jupiter, M_saturn, M_uranus, M_neptune, M_pluto]
-
-        elif model == 3:
-
-            spice.furnsh(os.path.join(SPICE_PATH, 'sb441-n16s.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000001.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000002.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000003.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000004.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000007.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000010.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000015.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000016.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000031.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000052.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000065.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000087.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000088.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000107.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000511.bsp'))
-            # spice.furnsh(os.path.join(SPICE_PATH, 'asteroids', '2000704.bsp'))
-
-            ceres = SpiceBody(spiceid='Ceres')
-            vesta = SpiceBody(spiceid='Vesta')
-            pallas = SpiceBody(spiceid='Pallas')
-            juno = SpiceBody(spiceid='2000003')
-            iris = SpiceBody(spiceid='2000007')
-            hygiea = SpiceBody(spiceid='2000010')
-            eunomia = SpiceBody(spiceid='2000015')
-            psyche = SpiceBody(spiceid='2000016')
-            euphrosyne = SpiceBody(spiceid='2000031')
-            europa = SpiceBody(spiceid='2000052')
-            cybele = SpiceBody(spiceid='2000065')
-            sylvia = SpiceBody(spiceid='2000087')
-            thisbe = SpiceBody(spiceid='2000088')
-            camilla = SpiceBody(spiceid='2000107')
-            davida = SpiceBody(spiceid='2000511')
-            interamnia = SpiceBody(spiceid='2000704')
-            
-            M_vesta = vesta.mass.value
-            M_ceres = ceres.mass.value
-            M_pallas = pallas.mass.value
-            M_juno = juno.mass.value
-            M_iris = iris.mass.value
-            M_hygiea = hygiea.mass.value
-            M_eunomia = eunomia.mass.value
-            M_psyche = psyche.mass.value
-            M_europa = europa.mass.value
-            M_cybele = cybele.mass.value
-            M_sylvia = sylvia.mass.value
-            M_thisbe = thisbe.mass.value
-            M_davida = davida.mass.value
-            M_interamnia = interamnia.mass.value
-            M_camilla = camilla.mass.value
-            M_euphrosyne = euphrosyne.mass.value
-
-            active_bodies = [sun, mercury, venus, earth, moon, mars, jupiter, saturn, uranus, neptune, pluto, 
-                             ceres, vesta, pallas, interamnia, juno, camilla, iris, hygiea, eunomia, psyche, 
-                             euphrosyne, europa, cybele, sylvia, thisbe, davida]
-            names = ['Sun', 'Mercury', 'Venus', 'Earth', 'Moon', 'Mars', 'Jupiter', 'Saturn', 'Uranus', 'Neptune', 'Pluto', 
-                     'Ceres', 'Vesta', 'Pallas', 'Interamnia', 'Juno', 'Camilla', 'Iris', 'Hygiea', 'Eunomia', 
-                     'Psyche', 'Euphrosyne', 'Europa', 'Cybele', 'Sylvia', 'Thisbe', 'Davida']
-            masses = [M_sun, M_mercury, M_venus, M_earth, M_moon, M_mars, M_jupiter, M_saturn, M_uranus, M_neptune, M_pluto, 
-                      M_ceres, M_vesta, M_pallas, M_interamnia, M_juno, M_camilla, M_iris, M_hygiea, M_eunomia, M_psyche, 
-                             M_euphrosyne, M_europa, M_cybele, M_sylvia, M_thisbe, M_davida]
-
-
-        else:
-            raise ValueError('Model not recognized. Check the documentation.')
-
-        startdate = Time(startdate, scale='tdb', format='jd')
-
-        bodies = [body.at(startdate) for body in active_bodies]
-
-        # create a dataframe of the massive bodies in the solar system
-        ss = pd.DataFrame()
-        ss['x'] = [body.x.au[0] for body in bodies]
-        ss['y'] = [body.y.au[0] for body in bodies]
-        ss['z'] = [body.z.au[0] for body in bodies]
-        ss['vx'] = [body.vx.value[0] for body in bodies]
-        ss['vy'] = [body.vy.value[0] for body in bodies]
-        ss['vz'] = [body.vz.value[0] for body in bodies]
-        ss['mass'] = masses
-        ss['name'] = names
-
-        sim = rebound.Simulation()
-        sim.units = ('day', 'AU', 'Msun')
-
-        for p in ss.itertuples():
-            sim.add(x=p.x, y=p.y, z=p.z,
-                    vx=p.vx, vy=p.vy, vz=p.vz,
-                    m=p.mass, hash=p.name)
-
-        sim.N_active = len(ss)
-
-        sim.testparticle_type = 0
-        sim.integrator = 'ias15'
-        #sim.ri_ias15.epsilon = 1e-12
-        
-
-        if gr == True:
-            import reboundx
-            rebx = reboundx.Extras(sim)
-            gr_force = rebx.load_force('gr_full')
-            gr_force.params["c"] = c.to(u.au / u.day).value
-            rebx.add_force(gr_force)
-        
-        return sim, names
-
     def analytic_propagate(self, epoch, units=Units()):
         '''
         analytically propagate all rocks to a common epoch
         '''
-        epoch = self.detect_timescale(np.atleast_1d(epoch), units.timescale)
+        if not isinstance(epoch, Time):
+            epoch = self.detect_timescale(np.atleast_1d(epoch), units.timescale)
         dt = (epoch.utc.jd - self.epoch.utc.jd) * u.day
         dM = self.n * dt
+
+        new_units = copy.deepcopy(self.units)
+        new_units.distance = u.au
+        new_units.angle = u.deg
         return self.__class__(a=self.a.au, 
                               e=self.e, 
                               inc=self.inc.deg, 
@@ -754,9 +574,11 @@ class SpaceRock(KeplerOrbit, Convenience):
                               arg=self.arg.deg, 
                               M=(self.M + dM).deg, 
                               name=self.name, 
+                              mass=self.mass.value,
                               epoch=np.repeat(epoch.utc.jd, len(self)), 
                               origin=self.origin, 
-                              frame=self.frame)
+                              frame=self.frame, 
+                              units=new_units)
         
 
     def orbits(self, N=1000):